+++ lib/functions/routes/get_app_pages.dart
@@ -1,11 +1,12 @@
 import 'dart:io';
 
+import 'package:get_cli/common/utils/pubspec/pubspec_utils.dart';
 import 'package:get_cli/core/structure.dart';
 import 'package:get_cli/functions/find_file/find_file_by_name.dart';
 import 'package:get_cli/samples/impl/get_app_pages.dart';
 import 'package:recase/recase.dart';
 
-Future<void> addAppPage(String name) async {
+Future<void> addAppPage(String name, String path) async {
   File appPagesFile = findFileByName('app_pages.dart');
   if (appPagesFile == null) {
     await AppPagesSample().create();
@@ -48,16 +49,18 @@ Future<void> addAppPage(String name) async {
       page:()=> ${namePascalCase}View(), 
       binding: ${namePascalCase}Binding(),
     ),''';
-  String import = Directory(Structure.replaceAsExpected(
-              path: Directory.current.path + '/lib/pages/'))
-          .existsSync()
-      ? 'pages'
-      : 'modules';
+
+  String import =
+      "import 'package:${await PubspecUtils.getProjectName()}/$path";
+
+  // String import = Directory(Structure.replaceAsExpected(
+  //             path: Directory.current.path + '/lib/pages/'))
+  //         .existsSync()
+  //     ? 'pages'
+  //     : 'modules';
   lines.insert(index, line);
-  lines.insert(0,
-      '''import '../$import/$nameSnakeCase/${nameSnakeCase}_binding.dart';''');
-  lines.insert(
-      0, '''import '../$import/$nameSnakeCase/${nameSnakeCase}_view.dart';''');
+  lines.insert(0, "$import'_binding.dart';");
+  lines.insert(0, "$import'_view.dart';");
 
   await appPagesFile.writeAsStringSync(lines.join('\n'));
 }